wire: make arena property values portable
In this commit, we generate transaction output values with the rapid int64 generator. The prior int generator used a bound larger than a 32-bit int, which prevented the wire tests from compiling on 32-bit targets even though the generated field itself is an int64.

+++ wire/arena_rapid_test.go
@@ -96,9 +96,7 @@ func genMsgTx(rt *rapid.T) *MsgTx {
 	numOut := rapid.IntRange(0, 6).Draw(rt, "numOut")
 	for i := 0; i < numOut; i++ {
 		tx.AddTxOut(&TxOut{
-			Value: int64(
-				rapid.IntRange(0, 1<<40).Draw(rt, "value"),
-			),
+			Value:    rapid.Int64Range(0, 1<<40).Draw(rt, "value"),
 			PkScript: genScript(rt, "pkScript", scriptMax),
 		})
 	}
